thesis_cv_quality_inspection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
pillow>=9.5.0
psutil>=5.9.0
tensorflow>=2.15.0
python-docx>=0.8.11
kaggle>=1.6.0
openpyxl>=3.1.2
